CDAP-9473 use api-spark-base as source
The spark and spark2 api modules will use the api-spark-base
module as source code in the dist profile, to ensure that the
base classes get compiled using the correct version of scala.

The base module is still treated as a module so that it is
recognized as code by IDEs, as IDEs will not allow multiple modules
to share the same source code. It is also a provided dependency
of the spark and spark2 api modules solely for IDE support.